--- src/form/image/ButtonImage.php
@@ -1,57 +1,0 @@
-<?php
-
-/*
- *
- *                            _____      _
- *     /\                    |  __ \    | |
- *    /  \   __ _ _   _  __ _| |__) |___| | __ _ _   _
- *   / /\ \ / _` | | | |/ _` |  _  // _ \ |/ _` | | | |
- *  / ____ \ (_| | |_| | (_| | | \ \  __/ | (_| | |_| |
- * /_/    \_\__, |\__,_|\__,_|_|  \_\___|_|\__,_|\__, |
- *               |_|                              |___/
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * @author AquaRelay Team
- * @link https://www.aquarelay.dev/
- *
- */
-
-declare(strict_types=1);
-
-namespace aquarelay\form\image;
-
-class ButtonImage
-{
-
-	private ImageType $imageType;
-
-	private string $data;
-
-	public function __construct(ImageType $imageType, string $data)
-	{
-		$this->imageType = $imageType;
-		$this->data = $data;
-	}
-
-	public function getImageType() : ImageType
-	{
-		return $this->imageType;
-	}
-
-	public function getData() : string
-	{
-		return $this->data;
-	}
-
-	public function toArray() : array
-	{
-		return [
-			"type" => $this->getImageType()->value,
-			"data" => $this->getData()
-		];
-	}
-}
